Refactor and remove generating JSON files
