+++ generator/src/main/kotlin/com/jaredrummler/androiddevicenames/Generator.kt
@@ -16,13 +16,19 @@
 
 package com.jaredrummler.androiddevicenames
 
-fun main() {
-  // Get the devices supported by Google Play
-  val devices = getDevices()
-  // Create the database
-  DatabaseGenerator(devices).generate()
-  // Generate all the JSON files
-  JsonGenerator(devices).generate()
-  // Create the method used in DeviceName.java.
-  MethodGenerator(devices).generate()
+class Generator {
+  companion object {
+    @JvmStatic
+    fun main(args: Array<String>) {
+      // Get the devices supported by Google Play
+      val devices = Devices.get()
+      // Create the database
+      DatabaseGenerator(devices).generate()
+      // Generate all the JSON files
+      // https://github.com/jaredrummler/AndroidDeviceNames/issues/31
+      // JsonGenerator(devices).generate()
+      // Create the method used in DeviceName.java.
+      MethodGenerator(devices).generate()
+    }
+  }
 }
